added blogs architecture, refactored code
