--- bottom_navigation_menu.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:id="@+id/membersFragment"
-        android:title="Members"
-        android:icon="@drawable/members"/>
-
-    <item android:id="@+id/eventsFragment"
-        android:title="Events"
-        android:icon="@drawable/events"/>
-
-    <item android:id="@+id/blogsFragment"
-        android:title="Blogs"
-        android:icon="@drawable/blogs"/>
-</menu>
